test: fix broken test for validate parameters
